fix sigmoid_focal_loss target type check bug

+++ mmcv/ops/focal_loss.py
@@ -36,7 +36,7 @@ class SigmoidFocalLossFunction(Function):
         if is_camb_parrots:
             assert isinstance(target, (torch.IntTensor, torch.cuda.IntTensor))
         else:
-            assert isinstance(target, (torch.IntTensor, torch.cuda.IntTensor))
+            assert isinstance(target, (torch.LongTensor, torch.cuda.LongTensor))
 
         assert input.dim() == 2
         assert target.dim() == 1
